troubleshooting a small issue with the money return system. should be fixed shortly.

+++ SodaMachine/Coins.cs
@@ -8,6 +8,6 @@ namespace SodaMachine
 {
     public class Coins
     {
-        public double value;
+        public decimal value;
     }
 }
